// lesson-one/lesson-one-experiment/drug.interface.ts
export interface Drug {
  drug_name: string;
  intervention_no: string;
  unk_no: string;
  drug_code: string;
  drug_recdose: string;
  drug_doserange: string;
  drug_maxdose: string;
  drug_mindose: string;
  drug_toxicdose: string;
  Drug_quantity_eliminated: string;
  drug_actionbp: string;
  drug_actionhr: string;
  drug_type: string;
  changeinbp: string;
  changeinhr: string;
  nature: string;
  group: string;
  drugDose: number;
  toxicLevel: number;
  druglevel: number;
  drugIncre:string;
}